Refactor ML4K API into separate class and improve HTTP calls
